move examples_manifest build script
