+++ crates/re_build_examples_manifest/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "re_build_examples_manifest"
+authors.workspace = true
+description = "Build the rerun web-viewer Wasm from source"
+edition.workspace = true
+homepage.workspace = true
+include.workspace = true
+license.workspace = true
+publish = true
+readme = "README.md"
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[package.metadata.docs.rs]
+all-features = true
+
+
+[dependencies]
+re_build_tools.workspace = true
+
+# External
+anyhow.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
+serde_yaml.workspace = true
