--- crates/re_viewer/data/examples_manifest.json
@@ -1,128 +1,0 @@
-[
-  {
-    "name": "arkit_scenes",
-    "title": "ARKit Scenes",
-    "description": "Visualize the ARKitScenes dataset, which contains color+depth images, the reconstructed mesh and labeled bounding boxes.",
-    "tags": [
-      "2D",
-      "3D",
-      "depth",
-      "mesh",
-      "object-detection",
-      "pinhole-camera"
-    ],
-    "demo_url": "https://demo.rerun.io/version/nightly/examples/arkit_scenes/",
-    "rrd_url": "https://demo.rerun.io/version/nightly/examples/arkit_scenes/data.rrd",
-    "thumbnail": {
-      "url": "https://static.rerun.io/arkit_scenes/fb9ec9e8d965369d39d51b17fc7fc5bae6be10cc/480w.png",
-      "width": 480,
-      "height": 243
-    }
-  },
-  {
-    "name": "detect_and_track_objects",
-    "title": "Detect and Track Objects",
-    "description": "Visualize object detection and segmentation using the Huggingface `transformers` library.",
-    "tags": [
-      "2D",
-      "huggingface",
-      "object-detection",
-      "object-tracking",
-      "opencv"
-    ],
-    "demo_url": "https://demo.rerun.io/version/nightly/examples/detect_and_track_objects/",
-    "rrd_url": "https://demo.rerun.io/version/nightly/examples/detect_and_track_objects/data.rrd",
-    "thumbnail": {
-      "url": "https://static.rerun.io/detect_and_track_objects/59f5b97a8724f9037353409ab3d0b7cb47d1544b/480w.png",
-      "width": 480,
-      "height": 279
-    }
-  },
-  {
-    "name": "dicom_mri",
-    "title": "Dicom MRI",
-    "description": "Example using a DICOM MRI scan. This demonstrates the flexible tensor slicing capabilities of the Rerun viewer.",
-    "tags": [
-      "tensor",
-      "mri",
-      "dicom"
-    ],
-    "demo_url": "https://demo.rerun.io/version/nightly/examples/dicom_mri/",
-    "rrd_url": "https://demo.rerun.io/version/nightly/examples/dicom_mri/data.rrd",
-    "thumbnail": {
-      "url": "https://static.rerun.io/dicom_mri/e39f34a1b1ddd101545007f43a61783e1d2e5f8e/480w.png",
-      "width": 480,
-      "height": 285
-    }
-  },
-  {
-    "name": "dna",
-    "title": "Helix",
-    "description": "Simple example of logging point and line primitives to draw a 3D helix.",
-    "tags": [
-      "3d",
-      "api-example"
-    ],
-    "demo_url": "https://demo.rerun.io/version/nightly/examples/dna/",
-    "rrd_url": "https://demo.rerun.io/version/nightly/examples/dna/data.rrd",
-    "thumbnail": {
-      "url": "https://static.rerun.io/helix/f4c375546fa9d24f7cd3a1a715ebf75b2978817a/480w.png",
-      "width": 480,
-      "height": 285
-    }
-  },
-  {
-    "name": "human_pose_tracking",
-    "title": "Human Pose Tracking",
-    "description": "Use the MediaPipe Pose solution to detect and track a human pose in video.",
-    "tags": [
-      "mediapipe",
-      "keypoint-detection",
-      "2D",
-      "3D"
-    ],
-    "demo_url": "https://demo.rerun.io/version/nightly/examples/human_pose_tracking/",
-    "rrd_url": "https://demo.rerun.io/version/nightly/examples/human_pose_tracking/data.rrd",
-    "thumbnail": {
-      "url": "https://static.rerun.io/human_pose_tracking/37d47fe7e3476513f9f58c38da515e2cd4a093f9/480w.png",
-      "width": 480,
-      "height": 272
-    }
-  },
-  {
-    "name": "plots",
-    "title": "Plots",
-    "description": "Demonstration of various plots and charts supported by Rerun.",
-    "tags": [
-      "2d",
-      "plots",
-      "api-example"
-    ],
-    "demo_url": "https://demo.rerun.io/version/nightly/examples/plots/",
-    "rrd_url": "https://demo.rerun.io/version/nightly/examples/plots/data.rrd",
-    "thumbnail": {
-      "url": "https://static.rerun.io/plots/c5b91cf0bf2eaf91c71d6cdcd4fe312d4aeac572/480w.png",
-      "width": 480,
-      "height": 271
-    }
-  },
-  {
-    "name": "structure_from_motion",
-    "title": "Structure from Motion",
-    "description": "Visualize a sparse reconstruction by COLMAP, a general-purpose Structure-from-Motion and Multi-View Stereo pipeline.",
-    "tags": [
-      "2D",
-      "3D",
-      "colmap",
-      "pinhole-camera",
-      "time-series"
-    ],
-    "demo_url": "https://demo.rerun.io/version/nightly/examples/structure_from_motion/",
-    "rrd_url": "https://demo.rerun.io/version/nightly/examples/structure_from_motion/data.rrd",
-    "thumbnail": {
-      "url": "https://static.rerun.io/structure_from_motion/b17f8824291fa1102a4dc2184d13c91f92d2279c/480w.png",
-      "width": 480,
-      "height": 275
-    }
-  }
-]
